+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.8.2 - TODO
 
 ### Fixed
+- Display Slice owner information (Issue [#409](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/409))
 - Non cloud facility ports should report local_name just like cloud facility ports (Issue [#402](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/402))
 - urllib3 deprecated warning (Issue [#397](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/397))
 - Improvements for get_interfaces and get_components (Issue [#378](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/378))